Add files via upload
KLtube.gendsp represents a single cylinder in the Kelly-Lochbaum vocal tract model. you can chain n tubes to construct the model, as shown in KL16poly.maxpat, where i have constructed a model with 16 tubes. it should be used with the poly~ object, with an argument of up 2, to upsample by a factor of 2.